+++ lib/src/platform.dart
@@ -1,6 +1,5 @@
-import 'package:universal_platform/universal_platform.dart';
+import 'package:flutter/foundation.dart';
 
 /// Check if the current platform is a mobile device.
-///
-/// Using the [UniversalPlatform] package,to cater for web and other platforms.
-bool get isMobileDevice => UniversalPlatform.isAndroid || UniversalPlatform.isIOS;
+bool get isMobileDevice =>
+    defaultTargetPlatform == TargetPlatform.android || defaultTargetPlatform == TargetPlatform.iOS;